Реализован на сокетах простейший HTTP 1.1 сервер. Добавил JavaDock-и. Переместил код в другой класс и разбил по методам
